# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ChartTag = typeof ChartTag[keyof typeof ChartTag];


export const ChartTag = {
  物量: '物量',
  ハネリズム: 'ハネリズム',
  階段: '階段',
  乱打: '乱打',
  トリル: 'トリル',
  同時押し: '同時押し',
  低速: '低速',
  高速: '高速',
  混合: '混合',
  未分類: '未分類',
} as const;

export interface CatalogSong {
  id: string;
  title: string;
  artist: string;
  chart: string;
  /** @nullable */
  officialDifficulty: number | null;
  unofficialDifficulty: number;
  pack: string;
  chartTags: ChartTag[];
  sourceUrl: string;
}

export type CatalogResponseSource = typeof CatalogResponseSource[keyof typeof CatalogResponseSource];


export const CatalogResponseSource = {
  'v-archive': 'v-archive',
  fallback: 'fallback',
} as const;

export interface CatalogResponse {
  songs: CatalogSong[];
  source: CatalogResponseSource;
  sourceUrl: string;
  fetchedAt: string;
}

export type SessionCreateInputButton = typeof SessionCreateInputButton[keyof typeof SessionCreateInputButton];


export const SessionCreateInputButton = {
  NUMBER_4: 4,
  NUMBER_5: 5,
  NUMBER_6: 6,
  NUMBER_8: 8,
} as const;

export interface SessionCreateInput {
  /**
     * @minLength 1
     * @maxLength 18
     */
  playerName: string;
  button: SessionCreateInputButton;
  /**
     * @minimum 1
     * @maximum 20
     */
  centerDifficulty: number;
  /**
     * @minimum 0
     * @maximum 5
     */
  lowerOffset: number;
  /**
     * @minimum 0
     * @maximum 5
     */
  upperOffset: number;
  p1Owned: string[];
  p2Owned: string[];
  selectedPacks: string[];
}

export interface SessionJoinInput {
  /**
     * @minLength 1
     * @maxLength 18
     */
  playerName: string;
}

export type SessionActionInputAction = typeof SessionActionInputAction[keyof typeof SessionActionInputAction];


export const SessionActionInputAction = {
  'set-setup': 'set-setup',
  'start-match': 'start-match',
  'draw-chart-tag': 'draw-chart-tag',
  'use-strategy': 'use-strategy',
  'resolve-strategy': 'resolve-strategy',
  'pick-song': 'pick-song',
  'advance-round': 'advance-round',
  'reset-match': 'reset-match',
} as const;

export type SessionActionInputPlayer = typeof SessionActionInputPlayer[keyof typeof SessionActionInputPlayer];


export const SessionActionInputPlayer = {
  p1: 'p1',
  p2: 'p2',
} as const;

export type SessionActionInputButton = typeof SessionActionInputButton[keyof typeof SessionActionInputButton];


export const SessionActionInputButton = {
  NUMBER_4: 4,
  NUMBER_5: 5,
  NUMBER_6: 6,
  NUMBER_8: 8,
} as const;

export interface SessionActionInput {
  action: SessionActionInputAction;
  player?: SessionActionInputPlayer;
  button?: SessionActionInputButton;
  /**
     * @minimum 1
     * @maximum 20
     */
  centerDifficulty?: number;
  /**
     * @minimum 0
     * @maximum 5
     */
  lowerOffset?: number;
  /**
     * @minimum 0
     * @maximum 5
     */
  upperOffset?: number;
  p1Owned?: string[];
  p2Owned?: string[];
  selectedPacks?: string[];
  selectedTag?: ChartTag;
  songId?: string;
}

export interface SessionRound {
  label: string;
  lower: number;
  upper: number;
  tag: ChartTag | null;
  p1Card: boolean;
  p2Card: boolean;
  resolved: boolean;
  /** @nullable */
  pickedSongId: string | null;
  /** @nullable */
  pickedSongTitle: string | null;
}

export type SessionStateRole = typeof SessionStateRole[keyof typeof SessionStateRole];


export const SessionStateRole = {
  host: 'host',
  guest: 'guest',
  full: 'full',
} as const;

export type SessionStateStatus = typeof SessionStateStatus[keyof typeof SessionStateStatus];


export const SessionStateStatus = {
  waiting: 'waiting',
  setup: 'setup',
  active: 'active',
  finished: 'finished',
} as const;

export interface SessionState {
  key: string;
  clientToken: string;
  role: SessionStateRole;
  status: SessionStateStatus;
  p1: string;
  /** @nullable */
  p2: string | null;
  button: number;
  centerDifficulty: number;
  lowerOffset: number;
  upperOffset: number;
  p1Owned: string[];
  p2Owned: string[];
  selectedPacks: string[];
  currentRound: number;
  rounds: SessionRound[];
  updatedAt: string;
}