# lib/api-zod/src/generated/types/catalogResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CatalogResponseSource } from './catalogResponseSource';
import type { CatalogSong } from './catalogSong';

export interface CatalogResponse {
  songs: CatalogSong[];
  source: CatalogResponseSource;
  sourceUrl: string;
  fetchedAt: string;
}