# lib/api-zod/src/generated/types/catalogSong.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChartTag } from './chartTag';

export interface CatalogSong {
  id: string;
  title: string;
  artist: string;
  chart: string;
  /** @nullable */
  officialDifficulty: number | null;
  unofficialDifficulty: number;
  pack: string;
  chartTags: ChartTag[];
  sourceUrl: string;
}