# lib/api-zod/src/generated/types/chartTag.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type ChartTag = typeof ChartTag[keyof typeof ChartTag];


export const ChartTag = {
  物量: '物量',
  ハネリズム: 'ハネリズム',
  階段: '階段',
  乱打: '乱打',
  トリル: 'トリル',
  同時押し: '同時押し',
  低速: '低速',
  高速: '高速',
  混合: '混合',
  未分類: '未分類',
} as const;