# lib/api-zod/src/generated/types/sessionActionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChartTag } from './chartTag';
import type { SessionActionInputAction } from './sessionActionInputAction';
import type { SessionActionInputButton } from './sessionActionInputButton';
import type { SessionActionInputPlayer } from './sessionActionInputPlayer';

export interface SessionActionInput {
  action: SessionActionInputAction;
  player?: SessionActionInputPlayer;
  button?: SessionActionInputButton;
  /**
     * @minimum 1
     * @maximum 20
     */
  centerDifficulty?: number;
  /**
     * @minimum 0
     * @maximum 5
     */
  lowerOffset?: number;
  /**
     * @minimum 0
     * @maximum 5
     */
  upperOffset?: number;
  p1Owned?: string[];
  p2Owned?: string[];
  selectedPacks?: string[];
  selectedTag?: ChartTag;
  songId?: string;
}