# lib/api-zod/src/generated/types/sessionCreateInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SessionCreateInputButton } from './sessionCreateInputButton';

export interface SessionCreateInput {
  /**
     * @minLength 1
     * @maxLength 18
     */
  playerName: string;
  button: SessionCreateInputButton;
  /**
     * @minimum 1
     * @maximum 20
     */
  centerDifficulty: number;
  /**
     * @minimum 0
     * @maximum 5
     */
  lowerOffset: number;
  /**
     * @minimum 0
     * @maximum 5
     */
  upperOffset: number;
  p1Owned: string[];
  p2Owned: string[];
  selectedPacks: string[];
}