# lib/api-zod/src/generated/types/sessionJoinInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SessionJoinInput {
  /**
     * @minLength 1
     * @maxLength 18
     */
  playerName: string;
}