# lib/api-zod/src/generated/types/sessionRound.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChartTag } from './chartTag';

export interface SessionRound {
  label: string;
  lower: number;
  upper: number;
  tag: ChartTag | null;
  p1Card: boolean;
  p2Card: boolean;
  resolved: boolean;
  /** @nullable */
  pickedSongId: string | null;
  /** @nullable */
  pickedSongTitle: string | null;
}