# lib/api-zod/src/generated/types/sessionState.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SessionRound } from './sessionRound';
import type { SessionStateRole } from './sessionStateRole';
import type { SessionStateStatus } from './sessionStateStatus';

export interface SessionState {
  key: string;
  clientToken: string;
  role: SessionStateRole;
  status: SessionStateStatus;
  p1: string;
  /** @nullable */
  p2: string | null;
  button: number;
  centerDifficulty: number;
  lowerOffset: number;
  upperOffset: number;
  p1Owned: string[];
  p2Owned: string[];
  selectedPacks: string[];
  currentRound: number;
  rounds: SessionRound[];
  updatedAt: string;
}